requirements.txt
Django==4.2
djangorestframewor==3.14.0
markdown-3.5.2
django-filter==23.5
djangorestframework-simplejwt==5.3.1
djangorestframework-simplejwt[crypto]
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.1.1
